[C8] Split serialize_work.go per work-entity kind (#2238)
model/feature/serialize_work.go was 878 LOC, 1.8x the 500-line ceiling. Split
the work-feature serialization per entity kind; the file also carried the
revolve and coil serialization, which are extracted to their own files so the
work core drops under the ceiling. Each file is now under the ceiling:

- serialize_work.go (228): the work-feature dispatch (MarshalWork/ApplyWork,
  serializeWorkFeature/restoreWorkFeature) and the shared reference encoding
  (workRefs/refStrings/unitSlice/p3Slice/point3From/unit3From).
- serialize_work_plane.go (162): work planes.
- serialize_work_axis.go (80): work axes.
- serialize_work_point.go (118): work points.
- serialize_revolve.go (220): revolve-feature serialization.
- serialize_coil.go (126): coil-feature serialization.

Behaviour-preserving: split programmatically by top-level declaration,
verified every non-comment code line is byte-identical (only import
redistribution differs). serialize_*.go is already funlen/gocyclo-excluded by
path in .golangci.yml (matches serialize_revolve/serialize_coil too), so no
config change is needed. Feature package builds, archguard passes, lint 0.
Net delta of recognizers / tolerance constants / fallback sites / type
assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2238.

(cherry picked from commit 21c9066cbda3efcc788c54f2137b3d93384a76f3)
