+++ model/feature/serialize_work.go
@@ -7,7 +7,6 @@ import (
 
 	"oblikovati.org/math"
 	"oblikovati.org/model/seq"
-	"oblikovati.org/model/sketch"
 )
 
 // This file serializes a part's USER work features (planes/axes/points) into the
@@ -90,83 +89,6 @@ func serializeWorkFeature(g *WorkGeometry, e userEntry) (WorkFeatureData, error)
 	}
 }
 
-func serializePlaneDef(def planeDefinition) (WorkFeatureData, error) {
-	d := WorkFeatureData{Collection: "plane", Kind: def.kindName(), Refs: refStrings(def.refs())}
-	switch v := def.(type) {
-	case *offsetPlaneDef:
-		d.Offset = v.distance() // persist the effective distance, including any browser edit
-	case *fixedFramePlaneDef:
-		p := v.origin()
-		d.Position = []float64{float64(p.X), float64(p.Y), float64(p.Z)}
-		d.XAxis, d.YAxis = unitSlice(v.x), unitSlice(v.y)
-	case *pointCloudFitPlaneDef:
-		// Persist the provenance link and the last good fit (the frozen frame), so the plane has
-		// geometry on load even before the cloud source is re-attached (#645).
-		d.CloudID = v.cloudID
-		d.Position = []float64{float64(v.origin.X), float64(v.origin.Y), float64(v.origin.Z)}
-		d.XAxis, d.YAxis = unitSlice(v.x), unitSlice(v.y)
-	case *linePlaneAnglePlaneDef:
-		d.Angle = v.angle()
-	case *twoPlanesPlaneDef:
-		if v.quadrant != nil { // persist the chosen bisector quadrant (#1844)
-			d.Solution = p3Slice(*v.quadrant)
-		}
-	case *planeAndTangentPlaneDef:
-		if v.proximity != nil { // persist the chosen tangent side (#1844)
-			d.Solution = p3Slice(*v.proximity)
-		}
-	case *lineAndTangentPlaneDef:
-		if v.proximity != nil {
-			d.Solution = p3Slice(*v.proximity)
-		}
-	case *threePointPlaneDef, *planeAndPointPlaneDef, *twoLinesPlaneDef,
-		*lineAndPointPlaneDef, *normalToCurvePlaneDef, *torusMidPlaneDef, *pointAndTangentPlaneDef:
-		// references only
-	default:
-		return WorkFeatureData{}, fmt.Errorf("no codec for work plane definition %q", def.kindName())
-	}
-	return d, nil
-}
-
-func serializeAxisDef(def axisDefinition) (WorkFeatureData, error) {
-	switch v := def.(type) {
-	case fixedAxisDef: // grounded "line" axis: persist its origin + direction (no references)
-		p := v.origin
-		return WorkFeatureData{
-			Collection: "axis", Kind: def.kindName(),
-			Position: []float64{float64(p.X), float64(p.Y), float64(p.Z)}, XAxis: unitSlice(v.dir),
-		}, nil
-	case twoPointsAxisDef, planeIntersectionAxisDef,
-		pointAndPlaneAxisDef, lineAndPointAxisDef, lineAndPlaneAxisDef, revolvedFaceAxisDef, edgeAxisDef:
-		return WorkFeatureData{Collection: "axis", Kind: def.kindName(), Refs: refStrings(def.refs())}, nil
-	default:
-		return WorkFeatureData{}, fmt.Errorf("no codec for work axis definition %q", def.kindName())
-	}
-}
-
-func serializePointDef(def pointDefinition) (WorkFeatureData, error) {
-	d := WorkFeatureData{Collection: "point", Kind: def.kindName(), Refs: refStrings(def.refs())}
-	switch v := def.(type) {
-	case positionPointDef:
-		p := v.at()
-		d.Position = []float64{float64(p.X), float64(p.Y), float64(p.Z)}
-	case *pointCloudPointDef:
-		p := v.FrozenPosition() // last good model position; the source re-derives it after relink (#645)
-		d.CloudID = v.cloudID
-		d.Position = []float64{float64(p.X), float64(p.Y), float64(p.Z)}
-	case curveEntityPointDef:
-		if v.proximity != nil { // record the solution-selection point so the side is stable (#1842)
-			d.Solution = []float64{float64(v.proximity.X), float64(v.proximity.Y), float64(v.proximity.Z)}
-		}
-	case planeAxisPointDef, pointRefPointDef, twoLinesPointDef, threePlanesPointDef, faceCenterPointDef,
-		edgeMidpointPointDef, centroidPointDef:
-		// references only
-	default:
-		return WorkFeatureData{}, fmt.Errorf("no codec for work point definition %q", def.kindName())
-	}
-	return d, nil
-}
-
 // ApplyWork rebuilds the user work features onto g (which already holds the origin
 // frame), in order, resolving each one's references as it goes.
 func ApplyWork(g *WorkGeometry, data []WorkFeatureData) error {
@@ -233,578 +155,6 @@ func restoreWorkFeature(g *WorkGeometry, d WorkFeatureData) error {
 	default:
 		return fmt.Errorf("unknown work collection %q", d.Collection)
 	}
-}
-
-// restorePlaneFeature rebuilds one user work plane from its recipe. Reference-only kinds
-// resolve through workRefs; fixed-frame/offset/angle kinds also carry scalar parameters,
-// re-installed as closures so a recompute re-reads them.
-func restorePlaneFeature(c *WorkPlanes, d WorkFeatureData) error {
-	switch d.Kind {
-	case "plane-offset":
-		return restoreRefPlane(d, 1, func(r []WorkRef) {
-			off := d.Offset
-			c.AddByPlaneAndOffset(r[0], func() float64 { return off })
-		})
-	case "three-points":
-		return restoreRefPlane(d, 3, func(r []WorkRef) { c.AddByThreePoints(r[0], r[1], r[2]) })
-	case "fixed-frame":
-		return restoreFixedFrame(c, d)
-	case "point-cloud-fit":
-		return restorePointCloudFit(c, d)
-	case "plane-point":
-		return restoreRefPlane(d, 2, func(r []WorkRef) { c.AddByPlaneAndPoint(r[0], r[1]) })
-	case "two-planes":
-		return restoreRefPlane(d, 2, func(r []WorkRef) {
-			if pt, ok := solutionPoint(d); ok {
-				c.AddByTwoPlanesToward(r[0], r[1], pt)
-				return
-			}
-			c.AddByTwoPlanes(r[0], r[1])
-		})
-	case "line-plane-angle":
-		return restoreRefPlane(d, 2, func(r []WorkRef) {
-			ang := d.Angle
-			c.AddByLinePlaneAndAngle(r[0], r[1], func() float64 { return ang })
-		})
-	case "two-lines":
-		return restoreRefPlane(d, 2, func(r []WorkRef) { c.AddByTwoLines(r[0], r[1]) })
-	case "line-point":
-		return restoreRefPlane(d, 2, func(r []WorkRef) { c.AddByLineAndPoint(r[0], r[1]) })
-	case "normal-to-curve":
-		return restoreRefPlane(d, 2, func(r []WorkRef) { c.AddByNormalToCurve(r[0], r[1]) })
-	case "torus-midplane":
-		return restoreRefPlane(d, 1, func(r []WorkRef) { c.AddByTorusMidPlane(r[0]) })
-	case "point-tangent":
-		return restoreRefPlane(d, 2, func(r []WorkRef) { c.AddByPointAndTangent(r[0], r[1]) })
-	case "plane-tangent":
-		return restoreRefPlane(d, 2, func(r []WorkRef) {
-			if pt, ok := solutionPoint(d); ok {
-				c.AddByPlaneAndTangentToward(r[0], r[1], pt)
-				return
-			}
-			c.AddByPlaneAndTangent(r[0], r[1])
-		})
-	case "line-tangent":
-		return restoreRefPlane(d, 2, func(r []WorkRef) {
-			if pt, ok := solutionPoint(d); ok {
-				c.AddByLineAndTangentToward(r[0], r[1], pt)
-				return
-			}
-			c.AddByLineAndTangent(r[0], r[1])
-		})
-	default:
-		return fmt.Errorf("no restore codec for work plane kind %q", d.Kind)
-	}
-}
-
-// restoreRefPlane resolves d's n references and calls add with them, centralizing the
-// arity check so each plane kind above stays a single line.
-func restoreRefPlane(d WorkFeatureData, n int, add func([]WorkRef)) error {
-	r, err := workRefs(d.Refs, n)
-	if err != nil {
-		return err
-	}
-	add(r)
-	return nil
-}
-
-// restoreFixedFrame rebuilds an AddFixed plane from its origin and two in-plane axes.
-func restoreFixedFrame(c *WorkPlanes, d WorkFeatureData) error {
-	origin, err := point3From(d.Position, "fixed-frame origin")
-	if err != nil {
-		return err
-	}
-	x, err := unit3From(d.XAxis, "fixed-frame X axis")
-	if err != nil {
-		return err
-	}
-	y, err := unit3From(d.YAxis, "fixed-frame Y axis")
-	if err != nil {
-		return err
-	}
-	c.AddFixed(func() math.Point3 { return origin }, x, y)
-	return nil
-}
-
-// restorePointCloudFit rebuilds an associative point-cloud-fit plane from its provenance id and
-// last good fit (the frozen frame). The live cloud source is re-attached after load by the host
-// (RelinkCloudFits), so until then the plane holds its frozen geometry (#645).
-func restorePointCloudFit(c *WorkPlanes, d WorkFeatureData) error {
-	origin, err := point3From(d.Position, "point-cloud-fit origin")
-	if err != nil {
-		return err
-	}
-	x, err := unit3From(d.XAxis, "point-cloud-fit X axis")
-	if err != nil {
-		return err
-	}
-	y, err := unit3From(d.YAxis, "point-cloud-fit Y axis")
-	if err != nil {
-		return err
-	}
-	c.addUser(&pointCloudFitPlaneDef{cloudID: d.CloudID, origin: origin, x: x, y: y, hasFit: true})
-	return nil
-}
-
-func restoreAxisFeature(c *WorkAxes, d WorkFeatureData) error {
-	if d.Kind == "line" { // grounded axis: rebuilt from its origin + direction, no references
-		return restoreLineAxis(c, d)
-	}
-	switch d.Kind { // single-reference axis kinds (a face or an edge) — #1840
-	case "revolved-face", "analytic-edge", "line-by-entity":
-		r, err := workRefs(d.Refs, 1)
-		if err != nil {
-			return err
-		}
-		switch d.Kind {
-		case "revolved-face":
-			c.AddByRevolvedFace(r[0])
-		case "analytic-edge":
-			c.AddByAnalyticEdge(r[0])
-		case "line-by-entity":
-			c.AddByLineByEntity(r[0])
-		}
-		return nil
-	}
-	r, err := workRefs(d.Refs, 2)
-	if err != nil {
-		return err
-	}
-	switch d.Kind {
-	case "two-points":
-		c.AddByTwoPoints(r[0], r[1])
-	case "plane-intersection":
-		c.AddByPlaneIntersection(r[0], r[1])
-	case "point-and-plane":
-		c.AddByPointAndPlane(r[0], r[1])
-	case "line-and-point":
-		c.AddByLineAndPoint(r[0], r[1])
-	case "line-and-plane":
-		c.AddByLineAndPlane(r[0], r[1])
-	default:
-		return fmt.Errorf("no restore codec for work axis kind %q", d.Kind)
-	}
-	return nil
-}
-
-// restoreLineAxis rebuilds a grounded "line" axis from its persisted origin + direction.
-func restoreLineAxis(c *WorkAxes, d WorkFeatureData) error {
-	o, err := point3From(d.Position, "line axis origin")
-	if err != nil {
-		return err
-	}
-	dir, err := unit3From(d.XAxis, "line axis direction")
-	if err != nil {
-		return err
-	}
-	c.AddByLine(o, dir)
-	return nil
-}
-
-func restorePointFeature(c *WorkPoints, d WorkFeatureData) error {
-	switch d.Kind {
-	case "position":
-		pos, err := point3From(d.Position, "position point")
-		if err != nil {
-			return err
-		}
-		c.AddByPosition(func() math.Point3 { return pos })
-		return nil
-	case "point-cloud-point":
-		pos, err := point3From(d.Position, "point-cloud point")
-		if err != nil {
-			return err
-		}
-		// The live cloud source is re-attached after load (RelinkCloudPoints); until then the point
-		// holds its frozen position (#645).
-		c.addUser(&pointCloudPointDef{cloudID: d.CloudID, frozen: pos, hasPos: true})
-		return nil
-	case "plane-axis-intersection":
-		r, err := workRefs(d.Refs, 2)
-		if err != nil {
-			return err
-		}
-		c.AddByPlaneAndAxisIntersection(r[0], r[1])
-		return nil
-	case "point":
-		r, err := workRefs(d.Refs, 1)
-		if err != nil {
-			return err
-		}
-		c.AddByPoint(r[0])
-		return nil
-	case "two-lines":
-		r, err := workRefs(d.Refs, 2)
-		if err != nil {
-			return err
-		}
-		c.AddByTwoLines(r[0], r[1])
-		return nil
-	case "three-planes":
-		r, err := workRefs(d.Refs, 3)
-		if err != nil {
-			return err
-		}
-		c.AddByThreePlanes(r[0], r[1], r[2])
-		return nil
-	case "face-center":
-		r, err := workRefs(d.Refs, 1)
-		if err != nil {
-			return err
-		}
-		c.AddByFaceCenter(r[0])
-		return nil
-	case "edge-midpoint":
-		r, err := workRefs(d.Refs, 1)
-		if err != nil {
-			return err
-		}
-		c.AddByMidpointOfEdge(r[0])
-		return nil
-	case "curve-and-entity":
-		r, err := workRefs(d.Refs, 2)
-		if err != nil {
-			return err
-		}
-		var prox *math.Point3
-		if p, ok := solutionPoint(d); ok {
-			prox = &p
-		}
-		c.AddByCurveAndEntity(r[0], r[1], prox)
-		return nil
-	case "centroid":
-		if len(d.Refs) == 0 {
-			return fmt.Errorf("centroid work point: no edge references")
-		}
-		c.AddAtCentroid(allWorkRefs(d.Refs)...)
-		return nil
-	default:
-		return fmt.Errorf("no restore codec for work point kind %q", d.Kind)
-	}
-}
-
-// RevolveData is a revolve's recipe: the sketch profile, the revolution axis (a
-// WorkRef, typically an origin axis or a user work axis), the swept angle, and the
-// boolean operation. Generation is still deferred, but the definition round-trips.
-type RevolveData struct {
-	Sketch     int     `yaml:"sketch"`
-	Profile    int     `yaml:"profile"`
-	Axis       string  `yaml:"axis,omitempty"`       // a work-axis WorkRef; empty ⇒ centerline mode
-	AxisSketch int     `yaml:"axisSketch,omitempty"` // 1-based sketch index of a specific centerline (0 = none)
-	AxisLine   int     `yaml:"axisLine,omitempty"`   // that centerline's line index
-	Angle      float64 `yaml:"angle,omitempty"`      // 0 ⇒ full revolution
-	Angle2     float64 `yaml:"angle2,omitempty"`     // second-direction sweep (#313)
-	Direction  string  `yaml:"direction,omitempty"`  // which way Angle sweeps (#2019); empty ⇒ positive
-	Operation  string  `yaml:"operation"`
-	// Extent is how the revolve terminates (#1860): empty ⇒ the angle extent (the default, so an
-	// ordinary revolve's recipe is unchanged), else "to-face", "from-to" or "to-next". ToPlane and
-	// FromPlane are the terminators' WorkRefs, named exactly as an extrude's are.
-	Extent    string `yaml:"extent,omitempty"`
-	ToPlane   string `yaml:"toPlane,omitempty"`
-	FromPlane string `yaml:"fromPlane,omitempty"`
-	// ProfilePoint selects the revolved region by an interior seed point (sketch 2-D cm) rather
-	// than by index, since an external author cannot predict the reader's region ordering. When
-	// present it wins over Profile; absent, the index is used unchanged.
-	ProfilePoint []float64 `yaml:"profilePoint,omitempty"`
-}
-
-func serializeRevolve(def *RevolveDefinition, sk SketchIndexer) (*RevolveData, error) {
-	idx, ok := sk.IndexOf(def.Sketch)
-	if !ok {
-		return nil, fmt.Errorf("revolve references a sketch that is not in the part")
-	}
-	op, err := operationName(def.Operation)
-	if err != nil {
-		return nil, err
-	}
-	d := &RevolveData{
-		Sketch: idx, Profile: def.ProfileIndex, Angle: evalFloat(def.Angle), Angle2: evalFloat(def.Angle2),
-		Direction: directionNames[def.Direction], Operation: op,
-		Extent: revolveExtentName(def.Extent), ToPlane: planeRefOf(def.ToPlane), FromPlane: planeRefOf(def.FromPlane),
-	}
-	switch {
-	case def.Axis != nil:
-		d.Axis = string(def.Axis.Key())
-	case def.AxisCenterline != nil: // a specific centerline (1-based so 0 means "none")
-		asi, ok := sk.IndexOf(def.AxisCenterlineSketch)
-		if !ok {
-			return nil, fmt.Errorf("revolve axis centerline references a sketch not in the part")
-		}
-		d.AxisSketch, d.AxisLine = asi+1, lineIndexOf(def.AxisCenterlineSketch, def.AxisCenterline)
-	}
-	return d, nil // both empty ⇒ revolve about the profile sketch's own centerline
-}
-
-// lineIndexOf returns the position of line within the sketch's line collection (-1 if absent).
-func lineIndexOf(sk *sketch.Sketch, line *sketch.Line) int {
-	for i := 0; i < sk.Lines().Count(); i++ {
-		if sk.Lines().Item(i) == line {
-			return i
-		}
-	}
-	return -1
-}
-
-func restoreRevolve(fs *PartFeatures, d *RevolveData, sk SketchIndexer, work *WorkGeometry) (*PartFeature, error) {
-	if d == nil {
-		return nil, fmt.Errorf("revolve feature is missing its payload")
-	}
-	pf, err := restoreRevolveAboutItsAxis(fs, d, sk, work)
-	if err != nil {
-		return nil, err
-	}
-	// The sweep (second angle, direction, extent) and the region seed are axis-independent, so they
-	// are applied once here rather than repeated down every axis branch.
-	if err := restoreRevolveExtent(pf, d, work); err != nil {
-		return nil, err
-	}
-	return withProfileSeed(restoreRevolveDirection(restoreSecondAngle(pf, d.Angle2), d.Direction), d.ProfilePoint), nil
-}
-
-// restoreRevolveExtent puts a persisted geometric termination back on a restored revolve (#1860).
-// An empty recipe extent is the angle extent, which needs no terminator.
-func restoreRevolveExtent(pf *PartFeature, d *RevolveData, work *WorkGeometry) error {
-	rf, ok := pf.feature.(*RevolveFeature)
-	if !ok || d.Extent == "" {
-		return nil
-	}
-	rf.def.Extent = parseExtentName(d.Extent)
-	var err error
-	if rf.def.ToPlane, err = restorePlaneRef(work, d.ToPlane); err != nil {
-		return err
-	}
-	rf.def.FromPlane, err = restorePlaneRef(work, d.FromPlane)
-	return err
-}
-
-// restorePlaneRef resolves an optional terminator reference; an empty one is simply absent.
-func restorePlaneRef(work *WorkGeometry, ref string) (*WorkPlane, error) {
-	if ref == "" {
-		return nil, nil
-	}
-	return resolvePlaneRef(work, ref)
-}
-
-// planeRefOf is the recipe spelling of an optional terminator plane.
-func planeRefOf(wp *WorkPlane) string {
-	if wp == nil {
-		return ""
-	}
-	return string(wp.Key())
-}
-
-// revolveExtentName is the recipe spelling of a revolve's termination. The angle extent is the
-// default and writes nothing, so an ordinary revolve's recipe is unchanged by #1860.
-func revolveExtentName(t ExtentType) string {
-	if t == DistanceExtent {
-		return ""
-	}
-	return extentNames[t]
-}
-
-// restoreRevolveAboutItsAxis rebuilds the revolve on whichever axis the recipe names: a specific
-// centerline, the profile sketch's own centerline, or a work axis.
-func restoreRevolveAboutItsAxis(fs *PartFeatures, d *RevolveData, sk SketchIndexer, work *WorkGeometry) (*PartFeature, error) {
-	skt, ok := sk.At(d.Sketch)
-	if !ok {
-		return nil, fmt.Errorf("revolve references sketch index %d, which does not exist", d.Sketch)
-	}
-	op, err := parseOperation(d.Operation)
-	if err != nil {
-		return nil, err
-	}
-	angle := d.Angle
-	// resolveSeed gives the fallback index; the seed is ALSO kept on the definition
-	// (withProfileSeed) so it re-resolves against the current regions every recompute — the
-	// sketch re-solves between load and recompute and reorders its regions (#region-seed).
-	profile := resolveSeed(skt, d.ProfilePoint, d.Profile)
-	if d.AxisSketch > 0 { // a specific centerline (1-based index)
-		line, err := restoredAxisCenterline(d, sk)
-		if err != nil {
-			return nil, err
-		}
-		return NewRevolveFeatures(fs).AddAboutCenterlineLine(skt, profile, line.sketch, line.line, func() float64 { return angle }, op), nil
-	}
-	if d.Axis == "" { // revolve about the profile sketch's own (single) centerline
-		return NewRevolveFeatures(fs).AddAboutCenterline(skt, profile, func() float64 { return angle }, op), nil
-	}
-	if work == nil {
-		return nil, fmt.Errorf("revolve needs the part's work geometry to resolve its axis")
-	}
-	axis, err := work.axis(WorkRef(d.Axis))
-	if err != nil {
-		return nil, fmt.Errorf("revolve axis: %w", err)
-	}
-	return NewRevolveFeatures(fs).Add(skt, profile, axis, func() float64 { return angle }, op), nil
-}
-
-// restoredCenterline is a revolve axis recovered from its 1-based sketch index and line index.
-type restoredCenterline struct {
-	sketch *sketch.Sketch
-	line   *sketch.Line
-}
-
-// restoredAxisCenterline resolves the recipe's axis centerline, reporting the offending index and
-// the range it had to fall in when it does not.
-func restoredAxisCenterline(d *RevolveData, sk SketchIndexer) (restoredCenterline, error) {
-	axisSk, ok := sk.At(d.AxisSketch - 1)
-	if !ok {
-		return restoredCenterline{}, fmt.Errorf("revolve axis centerline references sketch index %d, which does not exist", d.AxisSketch-1)
-	}
-	if d.AxisLine < 0 || d.AxisLine >= axisSk.Lines().Count() {
-		return restoredCenterline{}, fmt.Errorf("revolve axis centerline references line %d, outside the sketch's %d lines", d.AxisLine, axisSk.Lines().Count())
-	}
-	return restoredCenterline{sketch: axisSk, line: axisSk.Lines().Item(d.AxisLine)}, nil
-}
-
-// withProfileSeed records the interior seed point on a restored revolve so its region is
-// re-resolved by containment each recompute (survives the sketch re-solving — #region-seed).
-func withProfileSeed(pf *PartFeature, seed []float64) *PartFeature {
-	if len(seed) > 0 {
-		if rf, ok := pf.feature.(*RevolveFeature); ok {
-			rf.def.ProfileSeed = append([]float64(nil), seed...)
-		}
-	}
-	return pf
-}
-
-// restoreSecondAngle re-applies a persisted second-direction sweep onto a
-// freshly restored revolve (the centerline Add paths take one angle).
-func restoreSecondAngle(pf *PartFeature, angle2 float64) *PartFeature {
-	if angle2 <= 0 {
-		return pf
-	}
-	if rf, ok := pf.feature.(*RevolveFeature); ok {
-		rf.def.Angle2 = func() float64 { return angle2 }
-	}
-	return pf
-}
-
-// restoreRevolveDirection puts the saved sweep direction back on a restored revolve (#2019). It is
-// applied on EVERY axis path — a flipped revolve about a centerline is as real as one about a work
-// axis — so it sits outside the axis branches rather than inside one of them.
-func restoreRevolveDirection(pf *PartFeature, name string) *PartFeature {
-	if rf, ok := pf.feature.(*RevolveFeature); ok {
-		rf.def.Direction = parseDirectionName(name)
-	}
-	return pf
-}
-
-// CoilData is a coil's recipe: the sketch profile, the helix axis (a WorkRef), the
-// pitch (per revolution), the number of revolutions, the taper, and the operation.
-type CoilData struct {
-	Sketch      int     `yaml:"sketch"`
-	Profile     int     `yaml:"profile"`
-	Axis        string  `yaml:"axis"`
-	Pitch       float64 `yaml:"pitch,omitempty"`
-	Revolutions float64 `yaml:"revolutions,omitempty"`
-	Height      float64 `yaml:"height,omitempty"` // two-of-three shape spec (#316)
-	Taper       float64 `yaml:"taper,omitempty"`
-	Operation   string  `yaml:"operation"`
-	// Variable-pitch rail + end conditions (M06-F09, #624).
-	PitchRows []CoilPitchRowData `yaml:"pitchRows,omitempty"`
-	StartEnd  *CoilEndData       `yaml:"startEnd,omitempty"`
-	EndEnd    *CoilEndData       `yaml:"endEnd,omitempty"`
-	// Handedness and the flat-spiral flavour (#1883). LeftHanded is persisted as a flag rather
-	// than the enum so an existing document (no key) reads back as right-handed, the old default.
-	LeftHanded bool `yaml:"leftHanded,omitempty"`
-	Spiral     bool `yaml:"spiral,omitempty"`
-}
-
-// CoilPitchRowData is one persisted pitch station.
-type CoilPitchRowData struct {
-	Pitch      float64 `yaml:"pitch"`
-	Revolution float64 `yaml:"revolution"`
-}
-
-// CoilEndData is one persisted flat-end condition (radians).
-type CoilEndData struct {
-	TransitionAngle float64 `yaml:"transitionAngle,omitempty"`
-	FlatAngle       float64 `yaml:"flatAngle,omitempty"`
-}
-
-func serializeCoil(def *CoilDefinition, sk SketchIndexer) (*CoilData, error) {
-	idx, ok := sk.IndexOf(def.Sketch)
-	if !ok {
-		return nil, fmt.Errorf("coil references a sketch that is not in the part")
-	}
-	if def.Axis == nil {
-		return nil, fmt.Errorf("coil has no axis")
-	}
-	op, err := operationName(def.Operation)
-	if err != nil {
-		return nil, err
-	}
-	d := &CoilData{
-		Sketch: idx, Profile: def.ProfileIndex, Axis: string(def.Axis.Key()),
-		Pitch: evalFloat(def.Pitch), Revolutions: evalFloat(def.Revolutions),
-		Height: evalFloat(def.Height), Taper: def.Taper, Operation: op,
-		LeftHanded: def.Handedness == LeftHandedCoil, Spiral: def.Spiral,
-	}
-	for _, r := range def.PitchRows {
-		d.PitchRows = append(d.PitchRows, CoilPitchRowData(r))
-	}
-	d.StartEnd = coilEndData(def.StartEnd)
-	d.EndEnd = coilEndData(def.EndEnd)
-	return d, nil
-}
-
-// coilEndData persists a flat end (nil for natural).
-func coilEndData(c CoilEndCondition) *CoilEndData {
-	if !c.Flat {
-		return nil
-	}
-	return &CoilEndData{TransitionAngle: c.TransitionAngle, FlatAngle: c.FlatAngle}
-}
-
-func restoreCoil(fs *PartFeatures, d *CoilData, sk SketchIndexer, work *WorkGeometry) (*PartFeature, error) {
-	if d == nil {
-		return nil, fmt.Errorf("coil feature is missing its payload")
-	}
-	skt, ok := sk.At(d.Sketch)
-	if !ok {
-		return nil, fmt.Errorf("coil references sketch index %d, which does not exist", d.Sketch)
-	}
-	if work == nil {
-		return nil, fmt.Errorf("coil needs the part's work geometry to resolve its axis")
-	}
-	axis, err := work.axis(WorkRef(d.Axis))
-	if err != nil {
-		return nil, fmt.Errorf("coil axis: %w", err)
-	}
-	op, err := parseOperation(d.Operation)
-	if err != nil {
-		return nil, err
-	}
-	def := &CoilDefinition{
-		Sketch: skt, ProfileIndex: d.Profile, Axis: axis,
-		Pitch: constFloat(d.Pitch), Revolutions: constFloat(d.Revolutions),
-		Height: constFloat(d.Height), Taper: d.Taper, Operation: op,
-		Handedness: coilHandednessFromData(d.LeftHanded), Spiral: d.Spiral,
-	}
-	pf := NewCoilFeatures(fs).AddDefinition(def)
-	for _, r := range d.PitchRows {
-		def.PitchRows = append(def.PitchRows, CoilPitchRow(r))
-	}
-	def.StartEnd = coilEndFromData(d.StartEnd)
-	def.EndEnd = coilEndFromData(d.EndEnd)
-	return pf, nil
-}
-
-// coilHandednessFromData rebuilds the winding sense from the persisted flag (#1883).
-func coilHandednessFromData(left bool) CoilHandedness {
-	if left {
-		return LeftHandedCoil
-	}
-	return RightHandedCoil
-}
-
-// coilEndFromData rebuilds a persisted flat end (nil stays natural).
-func coilEndFromData(d *CoilEndData) CoilEndCondition {
-	if d == nil {
-		return CoilEndCondition{}
-	}
-	return CoilEndCondition{Flat: true, TransitionAngle: d.TransitionAngle, FlatAngle: d.FlatAngle}
 }
 
 // allWorkRefs casts every persisted reference string to a WorkRef — the variable-arity counterpart
